Safari-611.1.10/LayoutTests/fast/dom/HTMLButtonElement/null-handling.html

<DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests null handling of several HTMLButtonElement attributes");

var button = document.createElement("button");
shouldBeNull("button.getAttribute('formMethod')", "");
shouldBeNull("button.getAttribute('formEnctype')", "");
shouldBeNull("button.getAttribute('type')", "");
evalAndLog("button.formMethod = null");
shouldBeEqualToString("button.getAttribute('formMethod')", "null");
evalAndLog("button.formEnctype = null");
shouldBeEqualToString("button.getAttribute('formEnctype')", "null");
evalAndLog("button.type = null");
shouldBeEqualToString("button.getAttribute('type')", "null");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
